+++ packages/terra-menu/src/Menu.jsx
@@ -1,12 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Popup from 'terra-popup';
+import * as KeyCode from 'keycode-js';
 import classNames from 'classnames/bind';
 import MenuItem from './MenuItem';
 import MenuItemGroup from './MenuItemGroup';
 import MenuDivider from './MenuDivider';
 import MenuContent from './_MenuContent';
 import styles from './Menu.module.scss';
+import 'mutationobserver-shim';
+import './_contains-polyfill';
+import './_matches-polyfill';
 
 const cx = classNames.bind(styles);
 
@@ -53,12 +57,21 @@ const propTypes = {
    * Otherwise, the menu will display without an arrow and right aligned.
    */
   isArrowDisplayed: PropTypes.bool,
+  /**
+   * Allows assigning of root element custom data attribute for easy selecting of document base component.
+   */
+  rootSelector: PropTypes.string,
+  /**
+   * Allows assigning of trigger element custom data attribute for easy selecting of triggering component.
+   */
+  triggerElement: PropTypes.object,
 };
 
 const defaultProps = {
   isArrowDisplayed: false,
   isOpen: false,
   contentWidth: '240',
+  rootSelector: '#root',
 };
 
 class Menu extends React.Component {
@@ -67,7 +80,25 @@ class Menu extends React.Component {
     this.setPageDimensions = this.setPageDimensions.bind(this);
     this.push = this.push.bind(this);
     this.pop = this.pop.bind(this);
-    this.state = { stack: [this] };
+    this.disableContainerChildrenFocus = this.disableContainerChildrenFocus.bind(this);
+    this.enableContainerChildrenFocus = this.enableContainerChildrenFocus.bind(this);
+    this.menuElement = React.createRef();
+    this.state = {
+      stack: [this],
+    };
+  }
+
+  componentDidMount() {
+    // eslint-disable-next-line no-prototype-builtins
+    if (!Element.prototype.hasOwnProperty('inert')) {
+      // IE10 throws an error if wicg-inert is imported too early, as wicg-inert tries to set an observer on document.body which may not exist on import
+      // eslint-disable-next-line global-require
+      require('wicg-inert/dist/inert');
+    }
+    document.addEventListener('keydown', this.handleKeydown);
+    if (this.props.isOpen) {
+      this.disableContainerChildrenFocus();
+    }
   }
 
   componentDidUpdate(prevProps) {
@@ -75,6 +106,18 @@ class Menu extends React.Component {
       /* eslint-disable react/no-did-update-set-state */
       this.setState({ stack: [this] });
     }
+    if (this.props.isOpen) {
+      if (!prevProps.isOpen) {
+        this.disableContainerChildrenFocus();
+      }
+    } else if (prevProps.isOpen) {
+      this.enableContainerChildrenFocus();
+    }
+  }
+
+  componentWillUnmount() {
+    document.removeEventListener('keydown', this.handleKeydown);
+    this.enableContainerChildrenFocus();
   }
 
   setPageDimensions(node) {
@@ -87,6 +130,47 @@ class Menu extends React.Component {
       this.pageHeight = undefined;
       this.pageWidth = undefined;
     }
+  }
+
+  disableContainerChildrenFocus() {
+    const mainDocumentElement = document.querySelector(this.props.rootSelector);
+
+    if (mainDocumentElement) {
+      const inert = +mainDocumentElement.getAttribute('data-overlay-count');
+
+      if (!mainDocumentElement.hasAttribute('data-overlay-count')) {
+        mainDocumentElement.setAttribute('data-overlay-count', '1');
+        mainDocumentElement.setAttribute('inert', '');
+      } else if (mainDocumentElement && mainDocumentElement.hasAttribute('data-overlay-count')) {
+        mainDocumentElement.setAttribute('data-overlay-count', `${inert + 1}`);
+      }
+    }
+  }
+
+  enableContainerChildrenFocus() {
+    const mainDocumentElement = document.querySelector(this.props.rootSelector);
+
+    if (mainDocumentElement) {
+      const inert = +mainDocumentElement.getAttribute('data-overlay-count');
+
+      if (inert === 1) {
+        mainDocumentElement.removeAttribute('data-overlay-count');
+        mainDocumentElement.removeAttribute('inert');
+        if (this.props.triggerElement) {
+          // Shift focus back to element that was last focused prior to opening the menu
+          setTimeout(() => mainDocumentElement.removeAttribute('aria-hidden'), 0);
+        }
+      } else if (inert && inert > 1) {
+        mainDocumentElement.setAttribute('data-overlay-count', `${inert - 1}`);
+      }
+    }
+    setTimeout(() => {
+      if (this.props.triggerElement) {
+        // Shift focus back to element that was last focused prior to opening the modal
+        // eslint-disable-next-line no-console
+        this.props.triggerElement.focus();
+      }
+    }, 0); // Allows inert processing to finish before shifting focus back
   }
 
   pop() {
@@ -108,6 +192,18 @@ class Menu extends React.Component {
     });
   }
 
+  handleKeydown(e) {
+    const body = document.querySelector('body');
+    if (e.keyCode === KeyCode.KEY_ESCAPE && this.props.isOpen) {
+      if (this.menuElement.current) {
+        if (e.target === this.menuElement.current || this.menuElement.current.contains(e.target) || e.target === body) {
+          this.props.onRequestClose();
+        }
+      }
+    }
+  }
+
+
   render() {
     const {
       boundingRef,
@@ -120,6 +216,8 @@ class Menu extends React.Component {
       targetRef,
       isArrowDisplayed,
       contentWidth,
+      rootSelector,
+      triggerElement,
       ...customProps
     } = this.props;
     const arrowClass = cx([