+++ packages/terra-menu/src/terra-dev-site/test/menu/menu/DynamicMenuChildren.test.jsx
@@ -57,6 +57,8 @@ class DefaultMenu extends React.Component {
           isOpen={this.state.open}
           targetRef={this.getButtonNode}
           onRequestClose={this.handleRequestClose}
+          triggerElement={document.getElementById('default-button')}
+
         >
           {
             this.state.items.map(item => (