+++ packages/terra-menu/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 Updated `keycode-js` from `v1.0.4` to `v2.0.1`
 
+### Changed
+* Added inert/aria-hidden to prevent screen reader from moving focus out of popup.
+
 6.10.0 - (August 21, 2019)
 ------------------
 ### Changed